New superclasses for Integral, Sum, Product
The new classes work for all associative ops between limits.
The classes enable factor, expand, subs fixes to Sum and Integral
